+++ src/ResponseMapper.php
@@ -32,6 +32,7 @@
 
 namespace Wirecard\PaymentSdk;
 
+use Wirecard\PaymentSdk\Exception\MalformedResponseException;
 use Wirecard\PaymentSdk\Response\FailureResponse;
 use Wirecard\PaymentSdk\Response\FormInteractionResponse;
 use Wirecard\PaymentSdk\Response\InteractionResponse;
